Make logsystem replay cursors private
