+++ fdbserver/logsystem/LogSystemTypes.h
@@ -22,104 +22,7 @@
 #define FDBSERVER_LOGSYSTEM_LOGSYSTEMTYPES_H
 #pragma once
 
-#include "fdbrpc/Replication.h"
-#include "fdbserver/core/LogSystemConfig.h"
-#include "fdbserver/core/DBCoreState.h"
-
-struct ConnectionResetInfo : public ReferenceCounted<ConnectionResetInfo> {
-	double lastReset;
-	Future<Void> resetCheck;
-	int slowReplies;
-	int fastReplies;
-
-	ConnectionResetInfo() : lastReset(now()), resetCheck(Void()), slowReplies(0), fastReplies(0) {}
-};
-
-// Base cursor contract for consuming a sequential log peek stream.
-struct IPeekCursor {
-	virtual void setProtocolVersion(ProtocolVersion version) = 0;
-
-	virtual bool hasMessage() const = 0;
-	virtual VectorRef<Tag> getTags() const = 0;
-	virtual Arena& arena() = 0;
-	virtual ArenaReader* reader() = 0;
-	virtual StringRef getMessage() = 0;
-	virtual StringRef getMessageWithTags() = 0;
-	virtual void nextMessage() = 0;
-	virtual Future<Void> getMore(TaskPriority taskID = TaskPriority::TLogPeekReply) = 0;
-	virtual bool isExhausted() const = 0;
-	virtual const LogMessageVersion& version() const = 0;
-	virtual Version popped() const = 0;
-	virtual Version getMinKnownCommittedVersion() const = 0;
-	virtual void addref() = 0;
-	virtual void delref() = 0;
-};
-
-// Peek cursor that reports log location and can be cloned and repositioned for replay.
-struct IReplayPeekCursor : IPeekCursor {
-	// Upper bound on TLogPeekReply arenas retained before or while satisfying the next getMore().
-	virtual int64_t getMaxRetainedReplyCount() const = 0;
-	// Applies a per-reply cap before the cursor issues its first TLog peek. Zero leaves replies uncapped.
-	virtual void setReplyByteLimit(int limitBytes) = 0;
-	virtual Optional<UID> getPrimaryPeekLocation() const = 0;
-	virtual Optional<UID> getCurrentPeekLocation() const = 0;
-	virtual Version getMaxKnownVersion() const = 0;
-	virtual Reference<IReplayPeekCursor> cloneNoMore() = 0;
-	virtual void advanceTo(LogMessageVersion n) = 0;
-};
-
-class LogSet : NonCopyable, public ReferenceCounted<LogSet> {
-public:
-	std::vector<Reference<AsyncVar<OptionalInterface<TLogInterface>>>> logServers;
-	std::vector<Reference<AsyncVar<OptionalInterface<TLogInterface>>>> logRouters;
-	std::vector<Reference<AsyncVar<OptionalInterface<BackupInterface>>>> backupWorkers;
-	std::vector<Reference<ConnectionResetInfo>> connectionResetTrackers;
-	std::vector<Reference<Histogram>> tlogPushDistTrackers;
-	int32_t tLogWriteAntiQuorum;
-	int32_t tLogReplicationFactor;
-	std::vector<LocalityData> tLogLocalities;
-	TLogVersion tLogVersion;
-	Reference<IReplicationPolicy> tLogPolicy;
-	Reference<LocalitySet> logServerSet;
-	std::vector<int> logIndexArray;
-	std::vector<LocalityEntry> logEntryArray;
-	bool isLocal;
-	int8_t locality;
-	Version startVersion;
-	std::vector<Future<TLogLockResult>> replies;
-	std::vector<std::vector<int>> satelliteTagLocations;
-
-	LogSet()
-	  : tLogWriteAntiQuorum(0), tLogReplicationFactor(0), isLocal(true), locality(tagLocalityInvalid),
-	    startVersion(invalidVersion) {}
-	explicit LogSet(const TLogSet& tlogSet);
-	explicit LogSet(const CoreTLogSet& coreSet);
-
-	std::string logRouterString();
-	bool hasLogRouter(UID id) const;
-	bool hasBackupWorker(UID id) const;
-	std::string logServerString();
-	void populateSatelliteTagLocations(int logRouterTags,
-	                                   int oldLogRouterTags,
-	                                   int txsTags,
-	                                   int oldTxsTags,
-	                                   int cdcTags);
-	void checkSatelliteTagLocations();
-	int bestLocationFor(Tag tag);
-	void updateLocalitySet(std::vector<LocalityData> const& localities);
-	bool satisfiesPolicy(const std::vector<LocalityEntry>& locations);
-	void getPushLocations(
-	    VectorRef<Tag> tags,
-	    std::vector<int>& locations,
-	    int locationOffset,
-	    bool allLocations = false,
-	    const Optional<Reference<LocalitySet>>& restrictedLogSet = Optional<Reference<LocalitySet>>());
-
-private:
-	int satelliteTagLocationIndex(Tag tag) const;
-	std::vector<LocalityEntry> alsoServers, resultEntries;
-	std::vector<int> newLocations;
-};
+#include "fdbserver/logsystem/LogSystem.h"
 
 // Leaf replay cursor backed by a single TLog interface.
 class ServerPeekCursor final : public IReplayPeekCursor, public ReferenceCounted<ServerPeekCursor> {
