+++ fdbserver/logsystem/include/fdbserver/logsystem/LogSystem.h
@@ -42,6 +42,7 @@
 #include "fdbserver/core/SpanContextMessage.h"
 #include "fdbserver/core/TLogInterface.h"
 #include "fdbserver/core/WorkerInterface.actor.h"
+#include "fdbserver/logsystem/LogSet.h"
 #include "flow/Arena.h"
 #include "flow/Error.h"
 #include "flow/ActorCollection.h"
@@ -54,8 +55,39 @@ struct LogPushData;
 struct LocalityData;
 struct LogSystem;
 struct LogSystemConsumer;
-class LogSet;
-struct ConnectionResetInfo;
+
+// Base cursor contract for consuming a sequential log peek stream.
+struct IPeekCursor {
+	virtual void setProtocolVersion(ProtocolVersion version) = 0;
+
+	virtual bool hasMessage() const = 0;
+	virtual VectorRef<Tag> getTags() const = 0;
+	virtual Arena& arena() = 0;
+	virtual ArenaReader* reader() = 0;
+	virtual StringRef getMessage() = 0;
+	virtual StringRef getMessageWithTags() = 0;
+	virtual void nextMessage() = 0;
+	virtual Future<Void> getMore(TaskPriority taskID = TaskPriority::TLogPeekReply) = 0;
+	virtual bool isExhausted() const = 0;
+	virtual const LogMessageVersion& version() const = 0;
+	virtual Version popped() const = 0;
+	virtual Version getMinKnownCommittedVersion() const = 0;
+	virtual void addref() = 0;
+	virtual void delref() = 0;
+};
+
+// Peek cursor that reports log location and can be cloned and repositioned for replay.
+struct IReplayPeekCursor : IPeekCursor {
+	// Upper bound on TLogPeekReply arenas retained before or while satisfying the next getMore().
+	virtual int64_t getMaxRetainedReplyCount() const = 0;
+	// Applies a per-reply cap before the cursor issues its first TLog peek. Zero leaves replies uncapped.
+	virtual void setReplyByteLimit(int limitBytes) = 0;
+	virtual Optional<UID> getPrimaryPeekLocation() const = 0;
+	virtual Optional<UID> getCurrentPeekLocation() const = 0;
+	virtual Version getMaxKnownVersion() const = 0;
+	virtual Reference<IReplayPeekCursor> cloneNoMore() = 0;
+	virtual void advanceTo(LogMessageVersion n) = 0;
+};
 
 struct LogPushVersionSet {
 	Version prevVersion;
@@ -533,8 +565,6 @@ std::vector<T> LogSystem::getReadyNonError(std::vector<Future<T>> const& futures
 			result.push_back(f.get());
 	return result;
 }
-
-#include "LogSystemTypes.h"
 
 template <class T>
 OldLogData::OldLogData(const T& conf)
